Added: New base class KirchhoffLove for thin plate and shell elements
